+++ assigner/backends/__init__.py
@@ -1,5 +1,4 @@
 from assigner.backends.base import BackendBase, RepoError
 from assigner.backends.gitlab import GitlabBackend
 
-
-__all__ = [BackendBase, GitlabBackend, RepoError]
+pyflakes = [BackendBase, RepoError, GitlabBackend]